requirements.txt
fastapi[standard]
tortoise-orm[asyncmy]
pyjwt[crypto]
passlib[bcrypt]
pydantic_settings